pyproject.toml
[tool.poetry]
name = "rag-engine"
version = "0.1.0"
description = "RAG engine backend skeleton (Phase 1)"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.0"
PyJWT = "^2.8.0"
boto3 = "^1.34.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
alembic = "^1.17.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
black = "^24.4.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
